requirements.txt
pandas==2.2.3
langchain==0.3.7
langchain-openai==0.2.9
langgraph==0.2.5
python-dotenv==1.0.1
openpyxl==3.1.5
langchain-community==0.3.7
streamlit==1.40.1